+++ CHANGELOG.md
@@ -496,3 +496,6 @@
 * Configuration is now standardized instead of calling into static class functions or modifying objects directly.
 * Some parameters are now position sensitive in the parser call.  For example, putting 'linksto=Mediawiki|ignorecase=true|linksfrom=project' in that order would case 'linksto' to be case sensitive with 'linksfrom' being case insensitive.
 * Fixed several SQL injection exploits with 'ordercollation' and 'category'.
+* The 'dplcache' parameter was removed as it is now obselete.  Please use 'allowcachedresults' to control caching.
+* The 'dplcacheperiod' parameter was renamed to 'cacheperiod'.
+* The URL argument 'DPL_refresh' was renamed to 'cacherefresh'.